Rewrite action code to support batching of SQS messages
