+++ solution/backend/resources/urls.py
@@ -25,7 +25,7 @@ urlpatterns = [
     path("", ResourceViewSet.as_view({
         "get": "list",
     })),
-    path("<int:id>/edit", ResourceEditViewSet.as_view()),
+    path("<int:id>/edit", ResourceEditViewSet.as_view(), name="edit"),
     path("<int:id>/content", ContentTextViewSet.as_view(), name="content"),
     path("public", PublicResourceViewSet.as_view({
         "get": "list",
