+++ solution/backend/resources/utils/__init__.py
@@ -1,0 +1,2 @@
+from .aws_utils import *  # noqa
+from .general_purpose import *  # noqa
